Pass data to dialog
There were two problems:

* The project selected by the user was not being saved in the
controller. SOLUTION: Now, there is a variable named leiSelecionada
that stores this selected project. The assignment happens at the end of
the computeParcial function.

* When a dialog was shown a new controller instance was created. Since
the given name to the controller was SampleCtrl it gave a false
impression that the dialog would be using the already created
SampleCtrl controller. However, it was actually creating a new
controller, named SampleCtrl, inside the already created SampleCtrl.
Pretty much it was creating the SampleCtrl.SampleCtrl controller.
SOLUTION: In order to pass the data to the new controller, the locals
property in line 94 was used. This created a variable dataToPass that
should be given as a parameter to the function defined in line 95.